Capture the 'beforedeactivate' event in IE because the default behavior is for the active element to be deactivated whenever another element is clicked, regardless of whether that element belongs to a view that has 'acceptsFirstResponder' set to NO.
If we detect that the active element is “losing out” to an element that belongs to a view that does not accept keyPane or firstResponder, then cancel the event.  In this way, clients can create elements that behave as if they're part of a single user interface element — for example, a text field with a drop-down menu.  (Without this, clicking on a menu item element would cause the text field to lose focus!)